requirement.txt
aliyun-python-sdk-core==2.11.4
aliyun-python-sdk-core-v3==2.11.4
aliyun-python-sdk-kms==2.5.0
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.4.1
cached-property==1.5.1
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
crcmod==1.7
cryptography==2.4.2
cssselect==1.0.3
decorator==4.3.0
Django==1.11.5
django-autofixture==0.12.1
django-concurrency==1.4
django-configurations==2.1
django-debug-toolbar==1.11
django-elasticsearch-dsl==0.5.1
django-enumfields==1.0.0
django-extensions==1.8.1
django-filter==2.0.0
django-hashedfilenamestorage==2.3
django-jet==1.0.8
django-log-request-id==1.3.2
django-memoize==2.2.0
django-picklefield==1.0.0
django-storages==1.7.1
django-tables2==2.0.5
django-taggit==1.0.0
djangorestframework==3.9.0
docx==0.2.4
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
entrypoints==0.3
enum34==1.1.6
Faker==1.0.4
feedparser==5.2.1
flake8==3.7.7
freezegun==0.3.11
future==0.17.1
futures==3.1.1
-e git+https://github.com/livingbio/gchannel.git@a8896ed758bc3ac50f5609ba987a5fd14798b6f8#egg=gchannel
genv==0.1.121
gmediasearch==0.3.22
google-api-core==1.6.0
google-api-python-client==1.7.8
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud-core==0.29.0
google-cloud-language==1.1.1
google-cloud-storage==1.13.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
googletrans==2.4.0
gqueue==0.0.119
grpcio==1.17.1
grpcio-tools==1.17.1
h5py==2.9.0
hanziconv==0.3.2
httplib2==0.12.1
idna==2.6
ipaddress==1.0.22
isodate==0.6.0
itypes==1.1.0
jaraco.functools==1.20
jieba==0.39
Jinja2==2.10
jsonfield==1.0.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
livingbio-workspace==1545029605.705101
lxml==4.2.5
Markdown==2.6.11
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
multidict==4.5.2
nltk==3.3
numpy==1.15.2
oauthlib==3.0.1
olefile==0.46
oss2==2.6.0
paho-mqtt==1.4.0
pandas==0.23.4
pattern3==3.0.0
pbr==5.1.3
pdfminer.six==20181108
pdfminer3k==1.3.1
Pillow==5.3.0
pluggy==0.8.0
ply==3.11
portend==2.3
pprint==0.1
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.7.2
pyflakes==2.1.1
PyMySQL==0.9.2
pytest==4.0.1
pytest-cov==2.6.1
pytest-django==3.4.8
python-dateutil==2.7.5
python-slugify==1.2.6
python-twitter==3.5
pytz==2018.5
PyYAML==3.13
requests==2.18.4
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
schematics==2.1.0
scipy==1.2.0
selenium==3.141.0
simplejson==3.16.0
six==1.11.0
sortedcontainers==2.1.0
sqlparse==0.2.4
tempora==1.14
text-unidecode==1.2
ujson==1.35
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17.1
validators==0.12.3
vcrpy==2.0.1
virtualenv==16.1.0
wrapt==1.11.1
xmltodict==0.11.0
yarl==1.3.0
zc.lockfile==1.4
zhconv==1.4.0
